Small change to axios.ts nested ifs
When a function distributes amongst three separate conditions it can be difficult to reduce the code down. Switch statements could work - in this case it might not. For now I just flattened the nested structure. `if(!process.env.DEV)` isn't saying exactly the same thing as it did: `if(process.env.DEV)` however in this case there should be no change in behavior.